% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-nodes-to-graph-multi.R
\name{add_nodes_to_graph_multi}
\alias{add_nodes_to_graph_multi}
\title{Add nodes to a graph, allowing multiple points per edge}
\usage{
add_nodes_to_graph_multi(
  graph,
  xy,
  highway = "primary",
  wt_profile = NULL,
  wt_profile_file = NULL,
  debug = FALSE,
  dist_tol = 1e-06,
  intersections_only = FALSE,
  dist_min = dist_tol
)
}
\arguments{
\item{graph}{A dodgr street network data frame.}

\item{xy}{A two-column matrix or data frame of coordinates (x, y).}

\item{highway}{Character; highway type for new edges (see weight_streetnet).}

\item{wt_profile}{Character; weight profile to use for new edges (see weight_streetnet).}

\item{wt_profile_file}{Character; custom weight profile CSV file (see weight_streetnet).}

\item{debug}{Logical; if TRUE, print simple debug output.}

\item{dist_tol}{Numeric; tolerance for point matching.}

\item{intersections_only}{Logical; only split at intersections.}

\item{dist_min}{Numeric; minimum distance for point filtering.}
}
\value{
Modified graph with new nodes and edges.
}
\description{
This function extends \code{add_nodes_to_graph} by supporting multiple points on the same edge,
splitting edges as needed, and connecting every point to the closest (possibly new) vertex.
}
