% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare-heaps.R
\name{compare_heaps}
\alias{compare_heaps}
\title{compare_heaps}
\usage{
compare_heaps(graph, replications = 10)
}
\arguments{
\item{graph}{\code{data.frame} object representing the network graph}

\item{replications}{Number of replications to be used in comparison}
}
\value{
Result of \code{rbenachmar::benchmark} comparison in
\code{data.frame} form.
}
\description{
Perform timing comparison between different kinds of heaps as well as with
equivalent \code{igraph} routine \code{distances}.
}
