% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph-functions.R
\name{convert_graph}
\alias{convert_graph}
\title{convert_graph}
\usage{
convert_graph(graph, components = TRUE)
}
\arguments{
\item{graph}{A \code{data.frame} containing the edges of the graph}

\item{components}{If FALSE, components are not calculated (will generally
result in faster processing).}
}
\value{
A \code{data.frame} with the same number of rows but with columns of
\code{edge_id}, \code{from}, \code{to}, \code{d}, \code{w}, and
\code{component}, not all of which may be included.
}
\description{
Convert a graph represented as an arbitarily-structured \code{data.frame} to
standard 4 or 5-column format for submission to C++ routines
}
