% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dodgr_cache_off}
\alias{dodgr_cache_off}
\title{dodgr_cache_off}
\usage{
dodgr_cache_off()
}
\value{
Nothing; the function invisibly returns \code{TRUE} if successful.
}
\description{
Turn off all dodgr caching in current session. This is useful is speed is
paramount, and if graph contraction is not needed. Caching can be switched
back on with \link{dodgr_cache_on}.
}
\seealso{
Other cache: 
\code{\link{clear_dodgr_cache}()},
\code{\link{dodgr_cache_on}()}
}
\concept{cache}
