% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph-functions.R
\name{dodgr_compact_graph}
\alias{dodgr_compact_graph}
\title{dodgr_compact_graph}
\usage{
dodgr_compact_graph(graph, verts = NULL, quiet = TRUE)
}
\arguments{
\item{graph}{A flat table of graph edges. Must contain columns labelled
\code{from} and \code{to}, or \code{start} and \code{stop}. May also contain
similarly labelled columns of spatial coordinates (for example
\code{from_x}) or \code{stop_lon}).}

\item{verts}{Option \code{data.frame} of vertices obtained from
\code{dodgr_vertices} (submitting this will simply speed up conversion to
compact graph).}

\item{quiet}{If \code{FALSE}, display progress on screen}
}
\value{
A complex object with both the original graph and its compact verion
(\code{$original} and \code{$compact}, respectively), along with several
indexes used to map vertices and edges between the two.
}
\description{
Removes redundant (straight-line) vertices from graph, leaving only junction
vertices.
}
