% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hack.r
\name{dodgr_dists_pre}
\alias{dodgr_dists_pre}
\title{Calculate matrix of pair-wise distances between points.
Hacked par moi pour réduire les préchargements. Un graph "pré processé" est calculé
et utilisé à la place du graph
Ca réduit pas mal les coûts fixes.}
\usage{
dodgr_dists_pre(
  proc_g,
  from = NULL,
  to = NULL,
  shortest = TRUE,
  pairwise = FALSE,
  parallel = TRUE,
  quiet = TRUE,
  to_from_indices = NULL
)
}
\arguments{
\item{proc_g}{object representing the network preprocessed par process_graph
graph (see Notes)}

\item{from}{Vector or matrix of points \strong{from} which route distances are to
be calculated (see Notes)}

\item{to}{Vector or matrix of points \strong{to} which route distances are to be
calculated (see Notes)}

\item{shortest}{If \code{FALSE}, calculate distances along the \emph{fastest}
rather than shortest routes (see Notes).}

\item{pairwise}{If \code{TRUE}, calculate distances only between the ordered
pairs of \code{from} and \code{to}.}

\item{parallel}{If \code{TRUE}, perform routing calculation in parallel (see
details)}

\item{quiet}{If \code{FALSE}, display progress messages on screen.}

\item{to_from_indices}{(NULL) si non nul, remplace to/from et prend le racourci}

\item{heap}{Type of heap to use in priority queue. Options include
Fibonacci Heap (default; \code{FHeap}), Binary Heap (\code{BHeap}),
\verb{Trinomial Heap (}TriHeap\verb{), Extended Trinomial Heap (}TriHeapExt\verb{, and 2-3 Heap (}Heap23`).}
}
\value{
square matrix of distances between nodes
}
\description{
Calculate matrix of pair-wise distances between points.
Hacked par moi pour réduire les préchargements. Un graph "pré processé" est calculé
et utilisé à la place du graph
Ca réduit pas mal les coûts fixes.
}
\seealso{
Other distances: 
\code{\link{dodgr_distances}()},
\code{\link{dodgr_dists_categorical}()},
\code{\link{dodgr_dists_nearest}()},
\code{\link{dodgr_dists}()},
\code{\link{dodgr_flows_aggregate}()},
\code{\link{dodgr_flows_disperse}()},
\code{\link{dodgr_flows_si}()},
\code{\link{dodgr_isochrones}()},
\code{\link{dodgr_isodists}()},
\code{\link{dodgr_isoverts}()},
\code{\link{dodgr_paths}()},
\code{\link{dodgr_times}()}
}
\concept{distances}
