% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dists-proportional.R
\name{dodgr_dists_proportional}
\alias{dodgr_dists_proportional}
\title{Proportional distances along different edge categories}
\usage{
dodgr_dists_proportional(
  graph,
  from = NULL,
  to = NULL,
  heap = "BHeap",
  quiet = TRUE
)
}
\arguments{
\item{graph}{\code{data.frame} or equivalent object representing the network
graph which must have a column named "edge_type" which labels categories of
edge types along which proportional distances are to be aggregated (see
Note).}

\item{from}{Vector or matrix of points \strong{from} which route distances are to
be calculated (see Notes)}

\item{to}{Vector or matrix of points \strong{to} which route distances are to be
calculated (see Notes)}

\item{heap}{Type of heap to use in priority queue. Options include
Fibonacci Heap (default; \code{FHeap}), Binary Heap (\code{BHeap}),
\verb{Trinomial Heap (}TriHeap\verb{), Extended Trinomial Heap (}TriHeapExt\verb{, and 2-3 Heap (}Heap23`).}

\item{quiet}{If \code{FALSE}, display progress messages on screen.}
}
\value{
A list of distance matrices of equal dimensions (length(from),
length(to)), the first of which ("distance") holds the final distances, while
the rest are one matrix for each unique value of "edge_type", holding the
distances traversed along those types of edges only.
}
\description{
Proportional distances along different edge categories
}
\note{
The "edge_type" column in the graph can contain any kind of discrete or
categorical values, although integer values of 0 are not permissible. \code{NA}
values are ignored.
}
