% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph-functions.R
\name{dodgr_graph_components}
\alias{dodgr_graph_components}
\title{dodgr_graph_components}
\usage{
dodgr_graph_components(graph)
}
\arguments{
\item{graph}{A \code{data.frame} of edges}
}
\value{
Equivalent graph with additional \code{component} column,
sequentially numbered from 1 = largest component.
}
\description{
Identify connected components of graph and add corresponding \code{component}
column to \code{data.frame}.
}
