% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/break_long_edges.R
\name{dodgr_insert_vertex_parallel}
\alias{dodgr_insert_vertex_parallel}
\title{Insert a new vertex between two vertices in a graph}
\usage{
dodgr_insert_vertex_parallel(graph, v1, v2, x = NULL, y = NULL, cl = NULL)
}
\arguments{
\item{graph}{A flat table of graph edges. Must contain columns labelled
\code{from} and \code{to}, or \code{start} and \code{stop}. May also contain
similarly labelled columns of spatial coordinates (for example
\code{from_x}) or \code{stop_lon}).}

\item{v1}{Vertex defining start of graph edge along which new vertex is to be
inserted}

\item{v2}{Vertex defining end of graph edge along which new vertex is to be
inserted (order of \code{v1} and \code{v2} is not important).}

\item{x}{The \code{x}-coordinate of new vertex. If not specified, vertex is
created half-way between \code{v1} and \code{v2}.}

\item{y}{The \code{y}-coordinate of new vertex. If not specified, vertex is
created half-way between \code{v1} and \code{v2}.}

\item{cl}{Optional cluster object for parallel processing}
}
\value{
Modified version of graph with new vertex inserted between nominated vertices
}
\description{
Insert a new vertex between two vertices in a graph
}
\examples{
graph <- weight_streetnet (hampi)
e1 <- sample (nrow (graph), 1)
v1 <- graph$from_id [e1]
v2 <- graph$to_id [e1]
# insert new vertex in the middle of that randomly-selected edge:
graph2 <- dodgr_insert_vertex_parallel (graph, v1, v2)
nrow (graph)
nrow (graph2) # new edges added to graph2
}
