% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso.R
\name{dodgr_isodists}
\alias{dodgr_isodists}
\title{dodgr_isodists}
\usage{
dodgr_isodists(graph, from = NULL, dlim = NULL, heap = "BHeap")
}
\arguments{
\item{graph}{\code{data.frame} or equivalent object representing the network
graph (see Notes)}

\item{from}{Vector or matrix of points \strong{from} which isodistances are to
be calculated (see Notes)}

\item{dlim}{Desired limits of isodistances in metres.}

\item{heap}{Type of heap to use in priority queue. Options include
Fibonacci Heap (default; \code{FHeap}), Binary Heap (\code{BHeap}),
\code{Radix}, Trinomial Heap (\code{TriHeap}), Extended Trinomial Heap
(\code{TriHeapExt}, and 2-3 Heap (\code{Heap23}).}
}
\value{
A list of lists of isodistances as points sorted around the central
(origin) point. The first list of of \code{from} points, each member of which
has a list item for each specified value of \code{dlim}.
}
\description{
Isodistances from specified points. Function is fully vectorized to calculate
accept vectors of central points and vectors of isodistance thresholds.
}
\examples{
graph <- weight_streetnet (hampi)
from <- sample (graph$from_id, size = 100)
to <- sample (graph$to_id, size = 50)
d <- dodgr_dists (graph, from = from, to = to)
# d is a 100-by-50 matrix of distances between `from` and `to`
}
