% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph-functions.R
\name{dodgr_reinsert_verts}
\alias{dodgr_reinsert_verts}
\title{dodgr_reinsert_verts}
\usage{
dodgr_reinsert_verts(gc, verts = NULL)
}
\arguments{
\item{gc}{The result of call to \code{\link{dodgr_contract_graph}}, a list of
the three items of (i) the original graph; (ii) the contracted graph; and
(iii) an edge map related contracted to original edges.}

\item{verts}{List of vertices to be used for routing, or equivalent list of
geographical coordinates.}
}
\value{
A modified version of \code{gc}, with the contracted graph
(\code{$contracted}) including all vertices contained in \code{verts}.
}
\description{
Re-inserts vertices in a contracted graph. Vertices used for routing may often
be redundant - that is, be intermediate vertices between just two other
points - and will be removed in \code{\link{dodgr_contract_graph}}. This
function re-inserts vertices for routing prior to submitting graph to
\code{\link{dodgr_dists}}.
}
\examples{
graph <- weight_streetnet (hampi)
gc <- dodgr_contract_graph (graph)
nrow (gc$contracted) # size of contracted graph = 3,033
verts <- dodgr_vertices (graph)
verts <- verts [sample (nrow (verts), 100), ]
# gc <- dodgr_reinsert_verts (gc, verts = verts)
nrow (gc$contracted) # will be > 2,878
}
