% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths-and-flows.R
\name{dodgr_spatial_interaction}
\alias{dodgr_spatial_interaction}
\title{dodgr_spatial_interaction}
\usage{
dodgr_spatial_interaction(graph, nodes = NULL, dens = NULL, k = 2,
  contract = FALSE, heap = "BHeap", quiet = TRUE)
}
\arguments{
\item{graph}{\code{data.frame} or equivalent object representing the network
graph (see Details)}

\item{nodes}{Vector of points at which spatial interactions are to be
calculated (see Details)}

\item{dens}{Vector of corresponding densities used to calculate spatial
interactions}

\item{k}{Width coefficient of exponential spatial interaction model.}

\item{contract}{If \code{TRUE}, calculate flows on contracted graph before
mapping them back on to the original full graph (recommended as this will
generally be much faster).}

\item{heap}{Type of heap to use in priority queue. Options include
Fibonacci Heap (default; \code{FHeap}), Binary Heap (\code{BHeap}),
\code{Radix}, Trinomial Heap (\code{TriHeap}), Extended Trinomial Heap
(\code{TriHeapExt}, and 2-3 Heap (\code{Heap23}).}

\item{quiet}{If \code{FALSE}, display progress messages on screen.}
}
\value{
Matrix of same number of rows and columns as length of \code{nodes}
and \code{dens}, with rows containing the spatial interactions between each
node and all others.
}
\description{
Fit a single-constrained exponential spatial interaction model to a vector of
location densities.
}
