% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_streetnet.R
\name{dodgr_streetnet}
\alias{dodgr_streetnet}
\title{dodgr_streetnet}
\usage{
dodgr_streetnet(pts, expand = 0.05)
}
\arguments{
\item{pts}{List of points presumably containing spatial coordinates}

\item{expand}{Relative factor by which street network should extend beyond
limits defined by pts}
}
\value{
square matrix of distances between nodes
}
\description{
dodgr_streetnet
}
