% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph-conversion.R
\name{dodgr_to_igraph}
\alias{dodgr_to_igraph}
\title{dodgr_to_igraph}
\usage{
dodgr_to_igraph(graph)
}
\arguments{
\item{graph}{A \code{dodgr} graph}
}
\value{
The \code{igraph} equivalent of the input. Note that this will \emph{not}
be a dual-weighted graph.
}
\description{
Convert a \code{dodgr} graph to an \pkg{igraph}.
}
\examples{
graph <- weight_streetnet (hampi)
graphi <- dodgr_to_igraph (graph)
}
\seealso{
\link{igraph_to_dodgr}
}
