% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streetnet-fns.R
\name{dodgr_to_sf}
\alias{dodgr_to_sf}
\title{dodgr_to_sf}
\usage{
dodgr_to_sf(net, displ_progress = FALSE)
}
\arguments{
\item{net}{A \pkg{dodgr} network}

\item{displ_progress}{Should a progress bar be displayed?}
}
\value{
A Simple Features Collection (\code{sfc}) list of \code{LINESTRING}
objects.
}
\description{
Convert a \code{dodgr} graph as \code{data.frame} of all edges into a Simple
Features (\pkg{sf}) object through aggregating edges into \code{LINESTRING}
objects representing longest sequences between all junction nodes.
}
\note{
The output of this function corresponds to the edges obtained from
\code{dodgr_contract_graph}. An \pkg{sf} \code{data.frame} may be created by
appeding any data from the latter to the \code{sfc} output of this function -
see \pkg{sf} for details.
}
\examples{
hw <- weight_streetnet (hampi)
xy <- dodgr_to_sf (hw)
dim (hw) # 5,845 edges
length (xy) # 682 aggregated linestrings aggregated from those edges
}
