% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph-conversion.R
\name{dodgr_to_tidygraph}
\alias{dodgr_to_tidygraph}
\title{Convert a \code{dodgr} graph to an \pkg{tidygraph}.}
\usage{
dodgr_to_tidygraph(graph)
}
\arguments{
\item{graph}{A \code{dodgr} graph}
}
\value{
The \code{tidygraph} equivalent of the input
}
\description{
Convert a \code{dodgr} graph to an \pkg{tidygraph}.
}
\examples{
graph <- weight_streetnet (hampi)
grapht <- dodgr_to_tidygraph (graph)
}
\seealso{
Other conversion: 
\code{\link{dodgr_deduplicate_graph}()},
\code{\link{dodgr_to_igraph}()},
\code{\link{dodgr_to_sf}()},
\code{\link{dodgr_to_sfc}()},
\code{\link{igraph_to_dodgr}()}
}
\concept{conversion}
