% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-random-points.R
\name{generate_random_points}
\alias{generate_random_points}
\title{generate_random_points}
\usage{
generate_random_points(city, n)
}
\arguments{
\item{city}{Name of the city to be analysed}

\item{n}{Number of coordinates to be generated}
}
\value{
\code{matrix} with two columns of length n with random coordinates
}
\description{
Generate a numeric \code{matrix} of coordinates that lie within a given city.
The city can be specified by name.
}
