% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_streetnet.R
\name{get_streetnet}
\alias{get_streetnet}
\title{get_streetnet}
\usage{
get_streetnet(pts, expand = 0.05, txt = "from")
}
\arguments{
\item{pts}{List of points presumably containing spatial coordinates}

\item{expand}{Relative factor by which street network should extend beyond
limits defined by pts}

\item{txt}{Used for warning messages ("from" or "to")}
}
\value{
square matrix of distances between nodes
}
\description{
get_streetnet
}
