% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dodgr-package.r
\docType{data}
\name{hampi}
\alias{hampi}
\title{hampi}
\format{A Simple Features \code{sf} \code{data.frame} containing the street
network of Hampi.}
\description{
A sample street network from the township of Hampi, Karnataka, India.
}
\note{
Can be re-created with \code{get_streetnet("hampi india")}
}
\keyword{datasets}
