% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph-functions.R
\name{match_pts_to_graph}
\alias{match_pts_to_graph}
\title{match_pts_to_graph}
\usage{
match_pts_to_graph(verts, xy)
}
\arguments{
\item{verts}{A \code{data.frame} of vertices obtained from
\code{get_vertices(graph)}.}

\item{xy}{coordinates of points to be matched to the vertices}
}
\value{
A vector index into verts
}
\description{
Match spatial points to a spatial graph which contains vertex coordindates
}
