% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dists-proportional.R
\name{summary.dodgr_dists_proportional}
\alias{summary.dodgr_dists_proportional}
\title{Transform a result from 'dodgr_dists_proportional' to summary statistics}
\usage{
\method{summary}{dodgr_dists_proportional}(object, ...)
}
\arguments{
\item{object}{A 'dodgr_dists_proportional' object}
}
\value{
The summary statistics (invisibly)
}
\description{
Transform a result from 'dodgr_dists_proportional' to summary statistics
}
