% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{test}
\alias{test}
\title{test}
\usage{
test(graph, pq = "FHeap")
}
\arguments{
\item{graph}{\code{data.frame} object representing the network graph}

\item{pq}{Type of priority queue to use. Options include Fibonacci Heap
(default; \code{FHeap}), \code{Radix}, \code{Tri}, and \code{Heap23}.}

\item{quiet}{If FALSE, display progress information}
}
\value{
square matrix of distances between nodes
}
\description{
test
}
