% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_streetnet.R
\name{weight_streetnet}
\alias{weight_streetnet}
\title{weight_streetnet}
\usage{
weight_streetnet(graph, wt_profile = "bicycle")
}
\arguments{
\item{graph}{Street network extracted with \code{get_stretnet}}

\item{wt_profile}{Name of weighting profile}
}
\value{
A \code{data.frame} of edges representing the street network.
}
\description{
Weight (or re-weight) an OSM street network according to a named routino
profile, selected from (foot, horse, wheelchair, bicycle, moped, motorcycle,
motorcar, goods, hgv, psv).
}
