% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{weighting_profiles}
\alias{weighting_profiles}
\title{Weighting profiles}
\format{\code{data.frame} with profile names, means of transport and
weights.}
\usage{
weighting_profiles
}
\description{
Collection of weighting profiles used to adjust the routing process to
different means of transport. Original data taken from the Routino project.
}
\references{
\url{https://www.routino.org/xml/routino-profiles.xml}
}
\keyword{datasets}
