% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighting_profiles.R
\name{write_dodgr_wt_profile}
\alias{write_dodgr_wt_profile}
\title{write_dodgr_wt_profile}
\usage{
write_dodgr_wt_profile(file = NULL)
}
\arguments{
\item{file}{Full name (including path) of file to which to write. The \code{.json}
suffix will be automatically appended.}
}
\value{
TRUE if writing successful.
}
\description{
Write the \code{dodgr} street network weighting profiles to a local
\code{.json}-formatted file for manual editing and subsequent re-reading.
}
\seealso{
\link{weight_streetnet}

Other misc: 
\code{\link{compare_heaps}()},
\code{\link{dodgr_flowmap}()},
\code{\link{dodgr_full_cycles}()},
\code{\link{dodgr_fundamental_cycles}()},
\code{\link{dodgr_insert_vertex}()},
\code{\link{dodgr_sample}()},
\code{\link{dodgr_sflines_to_poly}()},
\code{\link{dodgr_vertices}()},
\code{\link{merge_directed_graph}()},
\code{\link{summary.dodgr_dists_categorical}()}
}
\concept{misc}
