% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{clear_dodgr_cache}
\alias{clear_dodgr_cache}
\title{Remove cached versions of \code{dodgr} graphs.}
\usage{
clear_dodgr_cache()
}
\value{
Nothing; the function silently clears any cached objects
}
\description{
This function should generally \emph{not} be needed, except if graph
structure has been directly modified other than through \code{dodgr} functions.
Graphs are cached based on a hash of the \code{edge_id}, \code{d}, \code{d_weighted},
\code{time}, and \code{time_weighted} columns, so manual changes to any of those will
be detected automatically. Manual changes to any other column or
attribute (for example, \code{from}/\code{to} vertex identifiers) will not be
reflected in the hash, and so will not necessarily be translated into
changes in \code{dodgr} output unless the cached versions are cleared using this
function. See
\url{https://github.com/UrbanAnalyst/dodgr/wiki/Caching-of-streetnets-and-contracted-graphs} # nolint
for details of caching process.
}
\examples{
clear_dodgr_cache ()
# Then call dodgr functions as usual:
graph <- weight_streetnet (hampi, wt_profile = "foot")
}
\seealso{
Other cache:
\code{\link[=dodgr_cache_off]{dodgr_cache_off()}},
\code{\link[=dodgr_cache_on]{dodgr_cache_on()}},
\code{\link[=dodgr_load_streetnet]{dodgr_load_streetnet()}},
\code{\link[=dodgr_save_streetnet]{dodgr_save_streetnet()}}
}
\concept{cache}
