% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clear_dodgr_cache}
\alias{clear_dodgr_cache}
\title{clear_dodgr_cache}
\usage{
clear_dodgr_cache()
}
\value{
Nothing; the function silently clears any cached objects
}
\description{
Remove cached versions of \code{dodgr} graphs. This function should generally
\emph{not} be needed, except if graph structure has been directly modified
other than through \code{dodgr} functions; for example by modifying edge weights
or distances. Graphs are cached based on the vector of edge IDs, so manual
changes to any other attributes will not necessarily be translated into
changes in \code{dodgr} output unless the cached versions are cleared using this
function. See
\url{https://github.com/ATFutures/dodgr/wiki/Caching-of-streetnets-and-contracted-graphs}
for details of caching process.
}
