% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare-heaps.R
\name{compare_heaps}
\alias{compare_heaps}
\title{compare_heaps}
\usage{
compare_heaps(graph, nverts = 100, replications = 2)
}
\arguments{
\item{graph}{\code{data.frame} object representing the network graph (or a
sub-sample selected with code{dodgr_sample})}

\item{nverts}{Number of vertices used to generate random sub-graph. If a
non-numeric value is given, the whole graph will be used.}

\item{replications}{Number of replications to be used in comparison}
}
\value{
Result of \code{rbenachmar::benchmark} comparison in
\code{data.frame} form.
}
\description{
Perform timing comparison between different kinds of heaps as well as with
equivalent \code{igraph} routine \code{distances}. To do this, a random
sub-graph containing a defined number of vertices is first selected.
Alternatively, this random sub-graph can be pre-generated with the
\code{dodgr_sample} function and passed directly.
}
\examples{
graph <- weight_streetnet (hampi)
compare_heaps (graph, nverts = 100, replications = 1)
}
