% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{dodgr_cache_off}
\alias{dodgr_cache_off}
\title{Turn off all dodgr caching in current session.}
\usage{
dodgr_cache_off()
}
\value{
Nothing; the function invisibly returns \code{TRUE} if successful.
}
\description{
This function is useful is speed is paramount, and if graph contraction is
not needed. Caching can be switched back on with \link{dodgr_cache_on}.
}
\seealso{
Other cache: 
\code{\link{clear_dodgr_cache}()},
\code{\link{dodgr_cache_on}()},
\code{\link{dodgr_load_streetnet}()},
\code{\link{dodgr_save_streetnet}()}
}
\concept{cache}
