% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{dodgr_cache_on}
\alias{dodgr_cache_on}
\title{dodgr_cache_on}
\usage{
dodgr_cache_on()
}
\value{
Nothing; the function invisibly returns \code{TRUE} if successful.
}
\description{
Turn on all dodgr caching in current session. This will only have an effect
after caching has been turned off with \link{dodgr_cache_off}.
}
\seealso{
Other cache: 
\code{\link{clear_dodgr_cache}()},
\code{\link{dodgr_cache_off}()},
\code{\link{dodgr_load_streetnet}()},
\code{\link{dodgr_save_streetnet}()}
}
\concept{cache}
