% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centrality.R
\name{dodgr_centrality}
\alias{dodgr_centrality}
\title{dodgr_centrality}
\usage{
dodgr_centrality(graph, contract = TRUE, edges = TRUE,
  heap = "BHeap")
}
\arguments{
\item{graph}{\code{data.frame} or equivalent object representing the network
graph (see Details)}

\item{contract}{If \code{TRUE}, centrality is calculated on contracted graph
before mapping back on to the original full graph. Note that for street
networks, in particular those obtained from the \pkg{osmdata} package, vertex
placement is effectively arbitrary except at junctions; centrality for such
graphs should only be calculated between the latter points, and thus
\code{contract} should always be \code{TRUE}.}

\item{edges}{If \code{TRUE}, centrality is calculated for graph edges, returning
the input \code{graph} with an additional \code{centrality} column; otherwise
centrality is calculated for vertices, returning the equivalent of
\code{dodgr_vertices(graph)}, with an additional vertex-based \code{centrality} column.}

\item{heap}{Type of heap to use in priority queue. Options include
Fibonacci Heap (default; \code{FHeap}), Binary Heap (\code{BHeap}),
\code{Radix}, Trinomial Heap (\code{TriHeap}), Extended Trinomial Heap
(\code{TriHeapExt}, and 2-3 Heap (\code{Heap23}).}
}
\value{
Modified version of graph with additonal \code{centrality} column added.
}
\description{
Calculate betweenness centrality for a \code{dodgr} network, in either vertex- or
edge-based form.
}
