% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph-functions.R
\name{dodgr_contract_graph}
\alias{dodgr_contract_graph}
\title{dodgr_contract_graph}
\usage{
dodgr_contract_graph(graph, verts = NULL, quiet = TRUE)
}
\arguments{
\item{graph}{A flat table of graph edges. Must contain columns labelled
\code{from} and \code{to}, or \code{start} and \code{stop}. May also contain
similarly labelled columns of spatial coordinates (for example
\code{from_x}) or \code{stop_lon}).}

\item{verts}{Option \code{data.frame} of vertices obtained from
\code{dodgr_vertices} (submitting this will simply speed up conversion to
compact graph).}

\item{quiet}{If \code{FALSE}, display progress on screen}
}
\value{
A list with the original graph (\code{$origina}), the contracted
graph (\code{$contracted}), and a map (\code{map}) between the IDs of edges
in the contracted graph and those in the original, full graph.
}
\description{
Removes redundant (straight-line) vertices from graph, leaving only junction
vertices.
}
\examples{
graph <- weight_streetnet (hampi)
nrow (graph) # 5,742
graph <- dodgr_contract_graph (graph)
nrow (graph$original) # size of contracted graph = 2,878
nrow (graph$map) # 3,692 = number of old edges replaced by new ones
}
