% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deduplicate.R
\name{dodgr_deduplicate_graph}
\alias{dodgr_deduplicate_graph}
\title{Deduplicate edges in a graph}
\usage{
dodgr_deduplicate_graph(graph)
}
\arguments{
\item{graph}{Any 'dodgr' graph or network.}
}
\value{
A potentially modified version of graph, with any formerly duplicated
edges reduces to single rows containing minimal weighted distances and times.
}
\description{
Graph may have duplicated edges, particularly when extracted as
\link{dodgr_streetnet} objects. This function de-duplicates any repeated
edges, reducing weighted distances and times to the minimal values from all
duplicates.
}
\seealso{
Other conversion: 
\code{\link{dodgr_to_igraph}()},
\code{\link{dodgr_to_sfc}()},
\code{\link{dodgr_to_sf}()},
\code{\link{dodgr_to_tidygraph}()},
\code{\link{igraph_to_dodgr}()}
}
\concept{conversion}
