% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dists-proportional.R
\name{dodgr_dists_proportional}
\alias{dodgr_dists_proportional}
\title{Proportional distances along different edge categories}
\usage{
dodgr_dists_proportional(
  graph,
  from = NULL,
  to = NULL,
  proportions_only = FALSE,
  dlimit = NULL,
  heap = "BHeap",
  quiet = TRUE
)
}
\arguments{
\item{graph}{\code{data.frame} or equivalent object representing the network
graph which must have a column named "edge_type" which labels categories of
edge types along which proportional distances are to be aggregated (see
Note).}

\item{from}{Vector or matrix of points \strong{from} which route distances are to
be calculated (see Notes)}

\item{to}{Vector or matrix of points \strong{to} which route distances are to be
calculated (see Notes)}

\item{proportions_only}{If \code{FALSE}, return full distance matrices for full
distances and for each edge category; if \code{TRUE}, return single vector of
proportional distances, like current \code{summary} function applied to full
results. See Note.}

\item{dlimit}{If \code{TRUE}, and no value to \code{to} is given, distances are
aggregated from each \code{from} point out to the specified distance limit (in
metres). The \code{proportions_only} argument has no effect when \code{dlimit = TRUE}.}

\item{heap}{Type of heap to use in priority queue. Options include
Fibonacci Heap (default; \code{FHeap}), Binary Heap (\code{BHeap}),
\verb{Trinomial Heap (}TriHeap\verb{), Extended Trinomial Heap (}TriHeapExt\verb{, and 2-3 Heap (}Heap23`).}

\item{quiet}{If \code{FALSE}, display progress messages on screen.}
}
\value{
If \code{dlimit = FALSE}, a list of distance matrices of equal dimensions (length(from),
length(to)), the first of which ("distance") holds the final distances, while
the rest are one matrix for each unique value of "edge_type", holding the
distances traversed along those types of edges only. If \code{dlimit = TRUE}, a
single matrix of total distances along all ways from each point, along with
distances along each of the different kinds of ways specified in the
"edge_type" column of the input graph.
}
\description{
Proportional distances along different edge categories
}
\note{
The "edge_type" column in the graph can contain any kind of discrete or
categorical values, although integer values of 0 are not permissible. \code{NA}
values are ignored. The function requires one full distance
matrix to be stored for each category of "edge_type" (unless
\code{proportions_only = TRUE}). It is wise to keep numbers of discrete types as
low as possible, especially for large distance matrices.

Setting the \code{proportions_only} flag to \code{TRUE} may be advantageous for
large jobs, because this avoids construction of the full matrices. This may
speed up calculations, but perhaps more importantly it may make possible
calculations which would otherwise require distance matrices too large to be
directly stored.
}
