% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowmap.R
\name{dodgr_flowmap}
\alias{dodgr_flowmap}
\title{dodgr_flowmap}
\usage{
dodgr_flowmap(net, filename, linescale = 5)
}
\arguments{
\item{net}{A street network with a \code{flow} column obtained from
\code{dodgr_flows}}

\item{filename}{Name of \code{.png} file to write to}

\item{linescale}{Maximal thickness of plotted lines}
}
\description{
Map the output of \link{dodgr_flows}
}
\note{
\code{net} should be first passed through \code{merge_directed_flows}
prior to plotting, otherwise lines for different directions will be overlaid.
}
