% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowmap.R
\name{dodgr_flowmap}
\alias{dodgr_flowmap}
\title{dodgr_flowmap}
\usage{
dodgr_flowmap(net, bbox = NULL, linescale = 1)
}
\arguments{
\item{net}{A street network with a \code{flow} column obtained from
\link{dodgr_flows_aggregate} or \link{dodgr_flows_disperse}}

\item{bbox}{If given, scale the map to this bbox, otherwise use entire extend
of \code{net}}

\item{linescale}{Maximal thickness of plotted lines}
}
\description{
Map the output of \link{dodgr_flows_aggregate} or \link{dodgr_flows_disperse}
}
\note{
\code{net} should be first passed through \code{merge_directed_flows}
prior to plotting, otherwise lines for different directions will be overlaid.
}
