% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{dodgr_paths_expand}
\alias{dodgr_paths_expand}
\title{Expand contracted paths back on to full, uncontracted graphs.}
\usage{
dodgr_paths_expand(paths, graph, graph_c, edge_map = NULL)
}
\arguments{
\item{paths}{A nested list of paths between pairs of (origin, destination)
points, generated by \link{dodgr_paths} on a graph previously contracted
with \link{dodgr_contract_graph}, so that \code{paths [[i]] [[j]]} is the path
between the i-th origin and j-th destination points. Each individual path
may be given either as a character vector of vertex IDs (as returned by
\link{dodgr_paths} with \code{vertices = TRUE}), or as an integer vector of row
indices into \code{graph_c}. Individual paths which are \code{NULL} or which trace
only a single vertex (with no transitions) are ignored.}

\item{graph}{The full, uncontracted graph from which \code{graph_c} was
generated with \link{dodgr_contract_graph}.}

\item{graph_c}{The contracted graph on which \code{paths} was calculated.}

\item{edge_map}{Optional edge map generated internally by
\link{dodgr_contract_graph}, and used to map contracted edges back on to
their corresponding sequences of edges in \code{graph}. If not given, this is
extracted directly from the dodgr cache.}
}
\value{
A nested list of the same structure as \code{paths}, in which
\code{result [[i]] [[j]]} is a \code{data.frame} of the rows of \code{graph}
corresponding to the full, uncontracted sequence of edges traced by
\code{paths [[i]] [[j]]}, ordered from start to end of that path. Entries for
which the corresponding input path is \code{NULL} or has no transitions are
returned as \code{NULL}.
}
\description{
Paths calculated with \link{dodgr_paths} on graphs contracted with
\link{dodgr_contract_graph} trace and return only the intervening junction
vertices. This function expands contracted paths back onto full sequences
of edges of original, uncontracted graphs.
}
\examples{
graph <- weight_streetnet (hampi)
graph_c <- dodgr_contract_graph (graph)
verts <- dodgr_vertices (graph_c)
from <- verts$id [1]
to <- verts$id [nrow (verts)]
paths_c <- dodgr_paths (graph_c, from = from, to = to)
paths <- dodgr_paths_expand (paths_c, graph, graph_c)
}
\seealso{
Other distances:
\code{\link[=dodgr_distances]{dodgr_distances()}},
\code{\link[=dodgr_dists]{dodgr_dists()}},
\code{\link[=dodgr_dists_categorical]{dodgr_dists_categorical()}},
\code{\link[=dodgr_dists_nearest]{dodgr_dists_nearest()}},
\code{\link[=dodgr_paths]{dodgr_paths()}},
\code{\link[=dodgr_times]{dodgr_times()}}
}
\concept{distances}
