% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dodgr-package.r
\docType{data}
\name{hampi}
\alias{hampi}
\title{hampi}
\format{
A Simple Features \code{sf} \code{data.frame} containing the street
network of Hampi.
}
\description{
A sample street network from the township of Hampi, Karnataka, India.
}
\note{
Can be re-created with the following command, which also removes
extraneous columns to reduce size:
}
\examples{
\dontrun{
hampi <- dodgr_streetnet("hampi india")
cols <- c ("osm_id", "highway", "oneway", "geometry")
hampi <- hampi [, which (names (hampi) \%in\% cols)]
}
# this 'sf data.frame' can be converted to a 'dodgr' network with
net <- weight_streetnet (hampi, wt_profile = 'foot')
}
\keyword{datasets}
