% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from-to-params.R
\name{process_graph}
\alias{process_graph}
\title{Préproccess le grpah pour gagner du temps}
\usage{
process_graph(graph, heap = "BHeap")
}
\arguments{
\item{graph}{le network}

\item{heap}{le heap}
}
\value{
un graph préprocéssé
}
\description{
Préproccess le grpah pour gagner du temps
}
