% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dists-categorical.R
\name{summary.dodgr_dists_categorical}
\alias{summary.dodgr_dists_categorical}
\title{Transform a result from 'dodgr_dists_categorical' to summary statistics}
\usage{
\method{summary}{dodgr_dists_categorical}(object, ...)
}
\arguments{
\item{object}{A 'dodgr_dists_categorical' object}

\item{...}{Extra parameters currently not used}
}
\value{
The summary statistics (invisibly)
}
\description{
Transform a result from 'dodgr_dists_categorical' to summary statistics
}
