% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{test}
\alias{test}
\title{test}
\usage{
test(graph, quiet = TRUE)
}
\arguments{
\item{graph}{\code{data.frame} object representing the network graph}

\item{quiet}{If FALSE, display progress information}
}
\value{
square matrix of distances between nodes
}
\description{
test
}
