% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dodgr-package.r
\docType{data}
\name{weighting_profiles}
\alias{weighting_profiles}
\title{weighting_profiles}
\format{\code{data.frame} with profile names, means of transport and
weights.}
\description{
Collection of weighting profiles used to adjust the routing process to
different means of transport. Modified from data taken from the Routino
project, with additional rows at end of table for waiting times in seconds at
traffic lights.
}
\references{
\url{https://www.routino.org/xml/routino-profiles.xml}
}
\keyword{datasets}
